fix up the tests from the mess I made via merging
