+++ validator_tests.py
@@ -13,6 +13,14 @@ class TestCSVValidator(unittest.TestCase):
         with self.assertRaises(NoDataError):
             validate_columns('test_files/tooFewColumns.csv')
 
+    def test_duplicated_column_names(self):
+        with self.assertRaises(DuplicateColumnError):
+            validate_columns('test_files/duplicatedColumnNames.csv')
+
+    def test_row_length_mismatch(self):
+        with self.assertRaises(ColumnMismatchError):
+            validate_columns('test_files/rowLengthMismatch.csv')
+
     def test_validate_columns_invalid_file(self):
         r = validate_columns('/this/better/not/exist/on/your/system')
         assert r is None
